refactor(workflows): remove duplicate activity-tree lookup and test doubles
BpmnDesigner.FindActivityById duplicated the Core JsonObjectActivityTreeExtensions.FindActivity
extension with identical semantics, so the Designer now calls the shared extension instead.
The BPMN editing test services also introduced a third near-identical ILocalizer/IActivityRegistry
stub; StateMachineActivityPickerDialogTests and StateMachinePresentationTests now share one
TestLocalizer/TestActivityRegistry pair from Support instead of keeping private copies.
