+++ src/modules/Elsa.Studio.Workflows.Tests/StateMachines/StateMachinePresentationTests.cs
@@ -4,9 +4,9 @@ using Elsa.Studio.Localization;
 using Elsa.Studio.Workflows.Designer;
 using Elsa.Studio.Workflows.Designer.Models;
 using Elsa.Studio.Workflows.DiagramDesigners.StateMachines.Presentation;
+using Elsa.Studio.Workflows.Tests.Support;
 using Microsoft.AspNetCore.Components.Web;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Localization;
 using Xunit;
 
 namespace Elsa.Studio.Workflows.Tests.StateMachines;
@@ -572,13 +572,5 @@ public sealed class StateMachinePresentationTests : BunitContext, IAsyncLifetime
             activity[name] = value;
 
         return activity;
-    }
-
-    private sealed class TestLocalizer : ILocalizer
-    {
-        public LocalizedString this[string? key] => new(key ?? string.Empty, key ?? string.Empty);
-
-        public LocalizedString this[string? key, params object[] arguments] =>
-            new(key ?? string.Empty, string.Format(key ?? string.Empty, arguments));
     }
 }
